moves state to logic
